+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/controller/RecipeFavoriteController.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.recipe.controller.favorite;
+package com.mars.app.domain.recipe.controller;
 
 import com.mars.app.domain.recipe.message.favorite.RecipeFavoriteMessagePublisher;
 import com.mars.common.dto.page.PageResponseDto;
@@ -7,7 +7,7 @@ import com.mars.app.aop.auth.AuthenticatedUser;
 import com.mars.app.component.auth.AuthenticationHolder;
 import com.mars.app.domain.recipe.dto.favorite.RecipeFavoriteListResponseDto;
 import com.mars.app.domain.recipe.dto.favorite.RecipeFavoriteResponseDto;
-import com.mars.app.domain.recipe.service.favorite.RecipeFavoriteService;
+import com.mars.app.domain.recipe.service.RecipeFavoriteService;
 import com.mars.common.dto.page.PageRequestVO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
