+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/repository/RecipeFavoriteRepository.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.recipe.repository.favorite;
+package com.mars.app.domain.recipe.repository;
 
 import com.mars.common.model.favorite.recipe.RecipeFavorite;
 import com.mars.common.model.recipe.Recipe;
