+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/repository/RecipeLikeRepository.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.recipe.repository.like;
+package com.mars.app.domain.recipe.repository;
 
 import com.mars.common.model.recipe.Recipe;
 import com.mars.common.model.recipe.RecipeLike;
