+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/service/RecipeCommentService.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.comment.recipe.service;
+package com.mars.app.domain.recipe.service;
 
 import static com.mars.common.exception.NPGExceptionType.NOT_FOUND_COMMENT;
 import static com.mars.common.exception.NPGExceptionType.NOT_FOUND_RECIPE;
@@ -6,11 +6,11 @@ import static com.mars.common.exception.NPGExceptionType.NOT_FOUND_USER;
 import static com.mars.common.exception.NPGExceptionType.UNAUTHORIZED_NO_AUTHENTICATION_CONTEXT;
 
 import com.mars.common.dto.page.PageResponseDto;
-import com.mars.app.domain.comment.recipe.dto.RecipeCommentRequestDto;
-import com.mars.app.domain.comment.recipe.dto.RecipeCommentResponseDto;
+import com.mars.app.domain.recipe.dto.comment.RecipeCommentRequestDto;
+import com.mars.app.domain.recipe.dto.comment.RecipeCommentResponseDto;
 import com.mars.common.dto.page.PageRequestVO;
 import com.mars.common.model.comment.recipe.RecipeComment;
-import com.mars.app.domain.comment.recipe.repository.RecipeCommentRepository;
+import com.mars.app.domain.recipe.repository.RecipeCommentRepository;
 import com.mars.common.model.recipe.Recipe;
 import com.mars.app.domain.recipe.repository.RecipeRepository;
 import com.mars.common.model.user.User;
