+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/service/RecipeEsService.java
@@ -7,11 +7,11 @@ import co.elastic.clients.elasticsearch._types.query_dsl.*;
 import co.elastic.clients.elasticsearch.core.SearchResponse;
 import com.mars.common.dto.page.PageRequestVO;
 import com.mars.common.exception.NPGExceptionType;
-import com.mars.app.domain.comment.recipe.repository.RecipeCommentRepository;
+import com.mars.app.domain.recipe.repository.RecipeCommentRepository;
 import com.mars.app.domain.recipe.builder.EsRecipeSearchQueryBuilder;
 import com.mars.app.domain.recipe.dto.RecipeEsResponseDto;
 import com.mars.common.model.recipe.RecipeEs;
-import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
+import com.mars.app.domain.recipe.repository.RecipeLikeRepository;
 import com.mars.common.model.recipe.RecipeLike;
 import com.mars.common.model.user.User;
 import java.util.ArrayList;
