+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/service/RecipeFavoriteService.java
@@ -1,12 +1,12 @@
-package com.mars.app.domain.recipe.service.favorite;
+package com.mars.app.domain.recipe.service;
 
-import com.mars.app.domain.comment.recipe.repository.RecipeCommentRepository;
-import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
+import com.mars.app.domain.recipe.repository.RecipeCommentRepository;
+import com.mars.app.domain.recipe.repository.RecipeLikeRepository;
 import com.mars.common.dto.page.PageResponseDto;
 import com.mars.common.dto.page.PageRequestVO;
 import com.mars.common.exception.NPGExceptionType;
 import com.mars.app.domain.recipe.dto.favorite.RecipeFavoriteListResponseDto;
-import com.mars.app.domain.recipe.repository.favorite.RecipeFavoriteRepository;
+import com.mars.app.domain.recipe.repository.RecipeFavoriteRepository;
 import com.mars.common.model.recipe.Recipe;
 import com.mars.common.model.recipe.RecipeLike;
 import com.mars.common.model.user.User;
