+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/user/service/UserService.java
@@ -5,12 +5,12 @@ import com.mars.app.domain.community.repository.CommunityRepository;
 import com.mars.common.dto.page.PageResponseDto;
 import com.mars.common.dto.page.PageRequestVO;
 import com.mars.common.exception.NPGExceptionType;
-import com.mars.app.domain.comment.recipe.dto.RecipeCommentResponseDto;
-import com.mars.app.domain.comment.recipe.repository.RecipeCommentRepository;
+import com.mars.app.domain.recipe.dto.comment.RecipeCommentResponseDto;
+import com.mars.app.domain.recipe.repository.RecipeCommentRepository;
 import com.mars.app.domain.recipe.dto.favorite.RecipeFavoriteListResponseDto;
-import com.mars.app.domain.recipe.repository.favorite.RecipeFavoriteRepository;
+import com.mars.app.domain.recipe.repository.RecipeFavoriteRepository;
 import com.mars.app.domain.recipe.dto.RecipeResponseDto;
-import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
+import com.mars.app.domain.recipe.repository.RecipeLikeRepository;
 import com.mars.app.domain.refrigerator.repository.RefrigeratorRepository;
 import com.mars.common.dto.user.MyPageDto;
 import com.mars.common.dto.user.UserInfoRequestDto;
