+++ NangPaGo-api/NangPaGo-app/src/test/java/com/mars/app/domain/recipe/message/RecipeLikeMessageConsumerTest.java
@@ -8,7 +8,7 @@ import static org.mockito.ArgumentMatchers.argThat;
 import com.mars.app.domain.recipe.dto.like.RecipeLikeMessageDto;
 import com.mars.app.domain.recipe.event.RecipeLikeEvent;
 import com.mars.app.domain.recipe.message.like.RecipeLikeMessageConsumer;
-import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
+import com.mars.app.domain.recipe.repository.RecipeLikeRepository;
 import com.mars.app.domain.recipe.repository.RecipeRepository;
 import com.mars.app.domain.user.repository.UserRepository;
 import com.mars.app.support.IntegrationTestSupport;
