+++ NangPaGo-api/NangPaGo-app/src/test/java/com/mars/app/domain/recipe/repository/RecipeLikeRepositoryTest.java
@@ -1,6 +1,5 @@
 package com.mars.app.domain.recipe.repository;
 
-import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
 import com.mars.common.model.recipe.RecipeLike;
 import com.mars.app.support.AbstractRepositoryTestSupport;
 import jakarta.transaction.Transactional;
