+++ NangPaGo-api/NangPaGo-app/src/test/java/com/mars/app/domain/user/service/UserServiceTest.java
@@ -3,7 +3,7 @@ package com.mars.app.domain.user.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.mars.app.domain.comment.recipe.dto.RecipeCommentResponseDto;
+import com.mars.app.domain.recipe.dto.comment.RecipeCommentResponseDto;
 import com.mars.app.domain.recipe.dto.favorite.RecipeFavoriteListResponseDto;
 import com.mars.app.domain.recipe.dto.RecipeResponseDto;
 import com.mars.app.domain.recipe.repository.RecipeRepository;
@@ -21,9 +21,9 @@ import com.mars.common.model.favorite.recipe.RecipeFavorite;
 import com.mars.common.model.recipe.Recipe;
 import com.mars.common.model.recipe.RecipeLike;
 import com.mars.common.model.user.User;
-import com.mars.app.domain.comment.recipe.repository.RecipeCommentRepository;
-import com.mars.app.domain.recipe.repository.favorite.RecipeFavoriteRepository;
-import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
+import com.mars.app.domain.recipe.repository.RecipeCommentRepository;
+import com.mars.app.domain.recipe.repository.RecipeFavoriteRepository;
+import com.mars.app.domain.recipe.repository.RecipeLikeRepository;
 import com.mars.app.domain.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import java.util.ArrayList;
